back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Alex Grigorovich <alex.grigorovich.mch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
debugpy = "1.8.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-lazy-fixture = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
src_paths = ["."]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
